fix: resolve all 24 mypy errors in src/ (24 -> 0)
Root causes, not suppressions where avoidable:
- internal_anchor_search.py: generate_simple_ratios() had a wrong return
  annotation (list[CandidateFormula] when it actually returns raw tuples,
  same convention as its 3 sibling generators) -- this alone cascaded
  into 11 of the 24 errors via search_anchor_formulas' tuple-unpacking loop.
- appendix_a1_procedure_registry.py: TableA1Row.h_mult/sigma_mult widened
  to float | None -- the z=0 anchor row in TABLE_A1 genuinely has no
  MULTING value, that's real source data, not a bug.
- cluster_schedule.py, k_a_independent.py, cluster_data_pipeline.py:
  float() wraps where a fractional-exponent ** or a stub-less third-party
  .value made mypy infer Any -- real numeric scalars at runtime.
- double_inversion_plots.py: imshow's extent= takes a tuple, was passed
  a list.
- report.py: scoped type: ignore[import-not-found] for a real, working
  sys.path.insert() runtime fallback import mypy can't trace statically.
- pyproject.toml: added an ignore_missing_imports override for
  astropy/astroquery, matching the existing scipy/matplotlib/pandas
  pattern.

ruff clean, 881/881 tests pass. No behavior changes.

+++ src/cluster_data_pipeline.py
@@ -125,7 +125,10 @@ def e_thermal_path_b(y500_arcmin2: float, z: float) -> float:
     y_sr = y500_arcmin2 * (np.pi / 10800.0) ** 2  # arcmin² → steradians
     d_a_m = Planck18.angular_diameter_distance(z).to(u.m).value
     e_joules = y_sr * d_a_m**2 * (_ME_KG * _C2 / _SIGMA_T)
-    return e_joules / (_MSUN_KG * _C2)
+    # WHY float(): astropy lacks type stubs, so .value is typed Any -- this
+    # is a real numeric scalar at runtime, float() makes the static type
+    # match the declared return type without changing behavior.
+    return float(e_joules / (_MSUN_KG * _C2))
 
 
 # ── Angular cross-match ─────────────────────────────────────────────────────
